created http.module.ts for CovalentHttpModule
